+++ eth_pydantic_types/address.py
@@ -43,7 +43,9 @@ class Address(HexStr20):
         )
 
     @classmethod
-    def __eth_pydantic_validate__(cls, value: Any, info: Optional[ValidationInfo] = None) -> str:
+    def __eth_pydantic_validate__(
+        cls, value: Any, info: Optional[ValidationInfo] = None
+    ) -> str:
         value = super().__eth_pydantic_validate__(value)
         return cls.to_checksum_address(value)
